+++ lib/phut/configuration.rb
@@ -45,7 +45,8 @@ module Phut
 
     def add_vhost(name, attrs)
       check_name_conflict name
-      @all[name] = Vhost.new(attrs[:ip], attrs[:promisc], name, @logger)
+      @all[name] =
+        Vhost.new(attrs[:ip], attrs[:mac], attrs[:promisc], name, @logger)
     end
 
     # This method smells of :reek:LongParameterList