+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Change log
 
 ## develop (unreleased)
+### New features
+* Add `mac` attribute to vhost.
 
 
 ## 0.2.0 (3/2/2015)